requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==45.0.7
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
stripe==12.5.1
twilio==9.8.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.10.0
yarl==1.20.1
